requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
requests==2.28.1
urllib3==1.26.11
Werkzeug==2.2.1
xmltodict==0.13.0
zipp==3.8.1
Flask-Cors==3.0.10
